+++ habapp_rules/actors/state_observer.py
@@ -22,13 +22,16 @@ CallbackType = typing.Callable[[EventTypes], None]
 class _StateObserverBase(HABApp.Rule, abc.ABC):
 	"""Base class for observer classes."""
 
-	def __init__(self, item_name: str, control_names: list[str] | None = None, group_names: list[str] | None = None):
+	def __init__(self, item_name: str, control_names: list[str] | None = None, group_names: list[str] | None = None, value_tolerance: int = 0):
 		"""Init state observer for switch item.
 
 		:param item_name: Name of observed item
 		:param control_names: list of control items.
 		:param group_names: list of group items where the item is a part of. Group item type must match with type of item_name
+		:param value_tolerance: used by all observers which handle numbers. It can be used to allow a difference when comparing new and old values.
 		"""
+		self._value_tolerance = value_tolerance
+
 		HABApp.Rule.__init__(self)
 		self._instance_logger = habapp_rules.core.logger.InstanceLogger(LOGGER, item_name)
 
@@ -122,6 +125,15 @@ class _StateObserverBase(HABApp.Rule, abc.ABC):
 		self._last_manual_event = event
 		callback: CallbackType = getattr(self, cb_name)
 		callback(event)
+
+	def _compare_values_with_tolerance(self, value_1: float, value_2: float) -> bool:
+		"""Compare values with tolerance
+
+		:param value_1: first value
+		:param value_2: second value
+		:return: true if values are the same (including the offset), false if not
+		"""
+		return abs((value_1 or 0) - (value_2 or 0)) > self._value_tolerance
 
 
 class StateObserverSwitch(_StateObserverBase):
@@ -218,7 +230,7 @@ class StateObserverDimmer(_StateObserverBase):
 			cb_brightness_change=callback_change)
 	"""
 
-	def __init__(self, item_name: str, cb_on: CallbackType, cb_off: CallbackType, cb_brightness_change: CallbackType | None = None, control_names: list[str] | None = None, group_names: list[str] | None = None) -> None:
+	def __init__(self, item_name: str, cb_on: CallbackType, cb_off: CallbackType, cb_brightness_change: CallbackType | None = None, control_names: list[str] | None = None, group_names: list[str] | None = None, value_tolerance: int = 0) -> None:
 		"""Init state observer for dimmer item.
 
 		:param item_name: Name of dimmer item
@@ -227,9 +239,10 @@ class StateObserverDimmer(_StateObserverBase):
 		:param cb_brightness_change: callback which is called if dimmer is on and brightness changed
 		:param control_names: list of control items. They are used to also respond to switch on/off via INCREASE/DECREASE
 		:param group_names: list of group items where the item is a part of. Group item type must match with type of item_name
+		:param value_tolerance: the tolerance can be used to allow a difference when comparing new and old values.
 		"""
 
-		_StateObserverBase.__init__(self, item_name, control_names, group_names)
+		_StateObserverBase.__init__(self, item_name, control_names, group_names, value_tolerance)
 
 		self._cb_on = cb_on
 		self._cb_off = cb_off
@@ -249,7 +262,7 @@ class StateObserverDimmer(_StateObserverBase):
 				self._value = 0
 				self._trigger_callback("_cb_off", event)
 
-			elif event.value != self._value:
+			elif self._compare_values_with_tolerance(event.value, self._value):
 				self._value = event.value
 				self._trigger_callback("_cb_brightness_change", event)
 
@@ -317,20 +330,22 @@ class StateObserverRollerShutter(_StateObserverBase):
 				)
 		"""
 
-	def __init__(self, item_name: str, cb_manual: CallbackType, control_names: list[str] | None = None, group_names: list[str] | None = None) -> None:
+	def __init__(self, item_name: str, cb_manual: CallbackType, control_names: list[str] | None = None, group_names: list[str] | None = None, value_tolerance: int = 0) -> None:
 		"""Init state observer for dimmer item.
 
 		:param item_name: Name of dimmer item
 		:param cb_manual: callback which is called if a manual interaction was detected
 		:param control_names: list of control items. They are used to also respond to switch on/off via INCREASE/DECREASE
 		:param group_names: list of group items where the item is a part of. Group item type must match with type of item_name
+		:param value_tolerance: the tolerance can be used to allow a difference when comparing new and old values.
 		"""
-		_StateObserverBase.__init__(self, item_name, control_names, group_names)
+		self._value_tolerance = value_tolerance
+		_StateObserverBase.__init__(self, item_name, control_names, group_names, value_tolerance)
 
 		self._cb_manual = cb_manual
 
 	def _check_manual(self, event: HABApp.openhab.events.ItemStateChangedEvent | HABApp.openhab.events.ItemCommandEvent) -> None:
-		if isinstance(event.value, (int, float)) and event.value != self._value:
+		if isinstance(event.value, (int, float)) and self._compare_values_with_tolerance(event.value, self._value):
 			self._value = event.value
 			self._trigger_callback("_cb_manual", event)
 
@@ -368,14 +383,15 @@ class StateObserverNumber(_StateObserverBase):
 	habapp_rules.actors.state_observer.StateObserverNumber("I01_01_Number", callback_value_changed)
 	"""
 
-	def __init__(self, item_name: str, cb_manual: CallbackType):
+	def __init__(self, item_name: str, cb_manual: CallbackType, value_tolerance: int = 0) -> None:
 		"""Init state observer for switch item.
 
 		:param item_name: Name of switch item
 		:param cb_manual: callback which should be called if manual change was detected
+		:param value_tolerance: the tolerance can be used to allow a difference when comparing new and old values.
 		"""
 		self._cb_manual = cb_manual
-		_StateObserverBase.__init__(self, item_name)
+		_StateObserverBase.__init__(self, item_name, value_tolerance=value_tolerance)
 
 	def _check_manual(self, event: HABApp.openhab.events.ItemStateChangedEvent | HABApp.openhab.events.ItemCommandEvent) -> None:
 		"""Check if light was triggered by a manual action
@@ -387,7 +403,7 @@ class StateObserverNumber(_StateObserverBase):
 			self._value = event.value
 			return
 
-		if event.value != self._value:
+		if self._compare_values_with_tolerance(event.value, self._value):
 			self._value = event.value
 			self._trigger_callback("_cb_manual", event)